// src/models/Dashboard.model.ts
export type DashbaordListResponse = {
  PanApplication: number;
  dscApplication: number;
  ItrApplication: number;
  gumastaApplication: number;
  msmeApplication: number;

};

export type DashbaordPendingListResponse = {
  totalpanapplications: number;
  totaldscapplications: number;
  totalitrapplications: number;
  totalgumastaapplications: number;
  totalmsmeapplications: number;

};